+++ examples/full.rs
@@ -5,13 +5,14 @@ use escpos::{driver::*, errors::Result};
 fn main() -> Result<()> {
     env_logger::init();
 
-    // let driver = NetworkDriver::open("192.168.1.248", 9100)?;
-    let driver = ConsoleDriver::open(true);
+    let driver = NetworkDriver::open("192.168.1.248", 9100)?;
+    // let driver = ConsoleDriver::open(true);
     let mut printer = Printer::new(driver, Protocol::default());
     printer
-        .debug_mode(Some(DebugMode::Hex))
+        .debug_mode(Some(DebugMode::Dec))
         .init()?
         .smoothing(true)?
+        .page_code(PageCode::default())?
         .bold(true)?
         .underline(UnderlineMode::Single)?
         .writeln("Bold underline")?