+++ CHANGELOG.md
@@ -18,6 +18,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add Select character code table command
+
 ### Changed
 
 ## `0.2.1` (2023-11-06) [CURRENT]